billingCardがほぼ完成
